+++ accommodation_requests/forms.py
@@ -879,10 +879,7 @@ class MoveGuestsConfirmationStep(forms.Form):
             destination = country.capitalize()
 
         # Set names based on the number of guests moved
-        names = moving_guests[0].get_full_name()
-        if len(moving_guests) > 1:
-            names_list = [guests.get_full_name() for guests in moving_guests]
-            names = ", ".join(names_list[:-1]) + f" and {names_list[-1]}"
+        names = MvAccommodationRequest.format_guest_names(moving_guests)
 
         headline = f"Are you sure you want to move {names} to {destination}?"