+++ unassigned_accommodation_requests/urls.py
@@ -3,7 +3,9 @@ from django.urls import path
 from .views import (
     ASSIGN_LOCAL_AUTHORITY_FORMS,
     AssignLocalAuthorityFormWizard,
+    HideUnassignedAccommodationRequestView,
     UnassignedAccommodationRequestsListView,
+    UnhideUnassignedAccommodationRequestView,
 )
 
 assign_local_authority_wizard = AssignLocalAuthorityFormWizard.as_view(
@@ -27,5 +29,15 @@ urlpatterns = [
         "<str:pk>/assign-local-authority",
         assign_local_authority_wizard,
         name="assign-local-authority",
+    ),
+    path(
+        "<str:pk>/hide/",
+        HideUnassignedAccommodationRequestView.as_view(),
+        name="hide",
+    ),
+    path(
+        "<str:pk>/unhide/",
+        UnhideUnassignedAccommodationRequestView.as_view(),
+        name="unhide",
     ),
 ]